requirements.txt
loguru
python-dotenv
selenium
webdriver-manager
python-telegram-bot==20.8
SQLAlchemy
